Move constants and exceptions in their own module
So I can work with constants outside of the gateway.
