+++ adyen/exceptions.py
@@ -1,0 +1,17 @@
+"""Adyen custom exceptions."""
+
+
+class MissingParameterException(ValueError):
+    pass
+
+
+class MissingFieldException(ValueError):
+    pass
+
+
+class UnexpectedFieldException(ValueError):
+    pass
+
+
+class InvalidTransactionException(ValueError):
+    pass
