+++ adyen/gateway.py
@@ -1,96 +1,17 @@
-# -*- coding: utf-8 -*-
-
 import base64
 import hashlib
 import hmac
 import logging
 
+from .constants import Constants
+from .exceptions import (
+    InvalidTransactionException,
+    MissingFieldException,
+    MissingParameterException,
+    UnexpectedFieldException,
+)
+
 logger = logging.getLogger('adyen')
-
-
-# ---[ CONSTANTS ]---
-
-class Constants:
-
-    ACCEPTED_NOTIFICATION = '[accepted]'
-
-    ACTION_URL = 'action_url'
-    ADYEN = 'adyen'
-    ALLOWED_METHODS = 'allowedMethods'
-    AUTH_RESULT = 'authResult'
-    BILLING_ADDRESS_TYPE = 'billingAddressType'
-    BLOCKED_METHODS = 'blockedMethods'
-    COUNTRY_CODE = 'countryCode'
-    CURRENCY = 'currency'
-    CURRENCY_CODE = 'currencyCode'
-    DELIVERY_ADDRESS_TYPE = 'deliveryAddressType'
-
-    EVENT_CODE = 'eventCode'
-    EVENT_CODE_AUTHORISATION = 'AUTHORISATION'
-    EVENT_DATE = 'eventDate'
-
-    FALSE = 'false'
-    IDENTIFIER = 'identifier'
-    LIVE = 'live'
-
-    MERCHANT_ACCOUNT = 'merchantAccount'
-    MERCHANT_ACCOUNT_CODE = 'merchantAccountCode'
-    MERCHANT_REFERENCE = 'merchantReference'
-    MERCHANT_RETURN_DATA = 'merchantReturnData'
-    MERCHANT_RETURN_URL = 'resURL'
-    MERCHANT_SIG = 'merchantSig'
-
-    OFFSET = 'offset'
-    OPERATIONS = 'operations'
-    ORIGINAL_REFERENCE = 'originalReference'
-
-    PAYMENT_AMOUNT = 'paymentAmount'
-    PAYMENT_METHOD = 'paymentMethod'
-    PAYMENT_RESULT_AUTHORISED = 'AUTHORISED'
-    PAYMENT_RESULT_REFUSED = 'REFUSED'
-    PAYMENT_RESULT_CANCELLED = 'CANCELLED'
-    PAYMENT_RESULT_PENDING = 'PENDING'
-    PAYMENT_RESULT_ERROR = 'ERROR'
-
-    PSP_REFERENCE = 'pspReference'
-    TEST_REFERENCE_PREFIX = 'test_AUTHORISATION'
-    REASON = 'reason'
-    RECURRING_CONTRACT = 'recurringContract'
-    SECRET_KEY = 'secret_key'
-    SEPARATOR = ':'
-    SESSION_VALIDITY = 'sessionValidity'
-    SKIN_CODE = 'skinCode'
-    SHIP_BEFORE_DATE = 'shipBeforeDate'
-    ADDITIONAL_DATA_PREFIX = 'additionalData.'
-
-    SHOPPER_EMAIL = 'shopperEmail'
-    SHOPPER_LOCALE = 'shopperLocale'
-    SHOPPER_REFERENCE = 'shopperReference'
-    SHOPPER_STATEMENT = 'shopperStatement'
-    SHOPPER_TYPE = 'shopperType'
-
-    SUCCESS = 'success'
-    TEST = 'test'
-    TRUE = 'true'
-    VALUE = 'value'
-
-
-# ---[ EXCEPTIONS ]---
-
-class MissingParameterException(ValueError):
-    pass
-
-
-class MissingFieldException(ValueError):
-    pass
-
-
-class UnexpectedFieldException(ValueError):
-    pass
-
-
-class InvalidTransactionException(ValueError):
-    pass
 
 
 # ---[ GATEWAY ]---
